requirements.txt
# Requirements for Google Cloud Run Video Processor + Render Deployment
flask>=3.0.0
gunicorn>=21.2.0
moviepy==1.0.3
pillow>=10.0.0
requests>=2.31.0
psycopg2-binary>=2.9.9
imageio-ffmpeg>=0.4.9
numpy
python-dotenv>=1.0.0
groq>=0.4.0
google-api-python-client>=2.100.0
